+++ src/UserInterface/src/MainWindow.cpp
@@ -501,7 +501,7 @@ void MainWindow::initUi() noexcept
             QMessageBox::information(
                 this,
                 QStringLiteral("Preview"),
-                QString::fromLatin1(
+                QStringLiteral(
                     "%1 is in a preview release phase: while it should be stable to use it is not "
                     "considered feature-complete.\n\n"
                     "This release v%2 \"%3\" introduces customisable shortcuts for MSFS "
@@ -676,8 +676,8 @@ void MainWindow::initViewUi() noexcept
 
 void MainWindow::initControlUi() noexcept
 {
-    ui->positionSlider->setRange(PositionSliderMin, PositionSliderMax);
-    ui->timestampTimeEdit->setDisplayFormat(QString::fromLatin1(TimestampFormat));
+    ui->positionSlider->setRange(::PositionSliderMin, ::PositionSliderMax);
+    ui->timestampTimeEdit->setDisplayFormat(QString::fromLatin1(::TimestampFormat));
 
     // Record/replay control buttons
     ui->recordButton->setAction(ui->recordAction);
@@ -705,34 +705,34 @@ void MainWindow::initReplaySpeedUi() noexcept
                                  new QAction(QStringLiteral("50 %"), this),
                                  new QAction(QStringLiteral("75 %"), this)};
     slowActions.at(0)->setShortcut(QKeySequence(Qt::CTRL | Qt::Key_1));
-    slowActions.at(0)->setProperty(ReplaySpeedProperty, Enum::underly(ReplaySpeed::Slow10));
+    slowActions.at(0)->setProperty(::ReplaySpeedProperty, Enum::underly(ReplaySpeed::Slow10));
     slowActions.at(1)->setShortcut(QKeySequence(Qt::CTRL | Qt::Key_2));
-    slowActions.at(1)->setProperty(ReplaySpeedProperty, Enum::underly(ReplaySpeed::Slow25));
+    slowActions.at(1)->setProperty(::ReplaySpeedProperty, Enum::underly(ReplaySpeed::Slow25));
     slowActions.at(2)->setShortcut(QKeySequence(Qt::CTRL | Qt::Key_3));
-    slowActions.at(2)->setProperty(ReplaySpeedProperty, Enum::underly(ReplaySpeed::Slow50));
+    slowActions.at(2)->setProperty(::ReplaySpeedProperty, Enum::underly(ReplaySpeed::Slow50));
     slowActions.at(3)->setShortcut(QKeySequence(Qt::CTRL | Qt::Key_4));
-    slowActions.at(3)->setProperty(ReplaySpeedProperty, Enum::underly(ReplaySpeed::Slow75));
+    slowActions.at(3)->setProperty(::ReplaySpeedProperty, Enum::underly(ReplaySpeed::Slow75));
 
     ui->normalSpeedAction->setCheckable(true);
     ui->normalSpeedAction->setShortcut(QKeySequence(Qt::SHIFT | Qt::Key_1));
-    ui->normalSpeedAction->setProperty(ReplaySpeedProperty, Enum::underly(ReplaySpeed::Normal));
+    ui->normalSpeedAction->setProperty(::ReplaySpeedProperty, Enum::underly(ReplaySpeed::Normal));
 
     QList<QAction *> fastActions{new QAction(QStringLiteral("2x"), this),
                                  new QAction(QStringLiteral("4x"), this),
                                  new QAction(QStringLiteral("8x"), this),
                                  new QAction(QStringLiteral("16x"), this)};
     fastActions.at(0)->setShortcut(QKeySequence(Qt::SHIFT | Qt::Key_2));
-    fastActions.at(0)->setProperty(ReplaySpeedProperty, Enum::underly(ReplaySpeed::Fast2x));
+    fastActions.at(0)->setProperty(::ReplaySpeedProperty, Enum::underly(ReplaySpeed::Fast2x));
     fastActions.at(1)->setShortcut(QKeySequence(Qt::SHIFT | Qt::Key_3));
-    fastActions.at(1)->setProperty(ReplaySpeedProperty, Enum::underly(ReplaySpeed::Fast4x));
+    fastActions.at(1)->setProperty(::ReplaySpeedProperty, Enum::underly(ReplaySpeed::Fast4x));
     fastActions.at(2)->setShortcut(QKeySequence(Qt::SHIFT | Qt::Key_4));
-    fastActions.at(2)->setProperty(ReplaySpeedProperty, Enum::underly(ReplaySpeed::Fast8x));
+    fastActions.at(2)->setProperty(::ReplaySpeedProperty, Enum::underly(ReplaySpeed::Fast8x));
     fastActions.at(3)->setShortcut(QKeySequence(Qt::SHIFT | Qt::Key_5));
-    fastActions.at(3)->setProperty(ReplaySpeedProperty, Enum::underly(ReplaySpeed::Fast16x));
+    fastActions.at(3)->setProperty(::ReplaySpeedProperty, Enum::underly(ReplaySpeed::Fast16x));
 
     ui->customSpeedAction->setCheckable(true);
     ui->customSpeedAction->setShortcut(QKeySequence(Qt::SHIFT | Qt::Key_6));
-    ui->customSpeedAction->setProperty(ReplaySpeedProperty, Enum::underly(ReplaySpeed::Custom));
+    ui->customSpeedAction->setProperty(::ReplaySpeedProperty, Enum::underly(ReplaySpeed::Custom));
 
     // Action group
     d->replaySpeedActionGroup = new QActionGroup(this);
@@ -817,7 +817,7 @@ void MainWindow::initReplaySpeedUi() noexcept
         normalSpeedRadioButton->setChecked(true);
     } else {
         d->customSpeedRadioButton ->setChecked(true);
-        d->customSpeedLineEdit->setText(d->unit.formatNumber(d->lastCustomReplaySpeedFactor, ReplaySpeedDecimalPlaces));
+        d->customSpeedLineEdit->setText(d->unit.formatNumber(d->lastCustomReplaySpeedFactor, ::ReplaySpeedDecimalPlaces));
     }
 
     d->replaySpeedUnitComboBox = new QComboBox(this);
@@ -961,15 +961,15 @@ void MainWindow::updateReplaySpeedUi() noexcept
 {
     if (d->customSpeedRadioButton->isChecked()) {
         d->customSpeedLineEdit->setEnabled(true);
-        d->customSpeedLineEdit->setText(d->unit.formatNumber(d->lastCustomReplaySpeedFactor, ReplaySpeedDecimalPlaces));
+        d->customSpeedLineEdit->setText(d->unit.formatNumber(d->lastCustomReplaySpeedFactor, ::ReplaySpeedDecimalPlaces));
 
         switch (Settings::getInstance().getReplaySpeeedUnit()) {
         case Replay::SpeedUnit::Absolute:
-            d->customSpeedLineEdit->setToolTip(tr("Custom replay speed factor in [%L1, %L2].").arg(ReplaySpeedAbsoluteMin).arg(ReplaySpeedAbsoluteMax));
+            d->customSpeedLineEdit->setToolTip(tr("Custom replay speed factor in [%L1, %L2].").arg(::ReplaySpeedAbsoluteMin).arg(::ReplaySpeedAbsoluteMax));
             d->customSpeedLineEdit->setValidator(d->customReplaySpeedFactorValidator);
             break;
         case Replay::SpeedUnit::Percent:
-            d->customSpeedLineEdit->setToolTip(tr("Custom replay speed % in [%L1%, %L2%].").arg(ReplaySpeedAbsoluteMin * 100.0).arg(ReplaySpeedAbsoluteMax * 100.0));
+            d->customSpeedLineEdit->setToolTip(tr("Custom replay speed % in [%L1%, %L2%].").arg(::ReplaySpeedAbsoluteMin * 100.0).arg(::ReplaySpeedAbsoluteMax * 100.0));
             d->customSpeedLineEdit->setValidator(d->customReplaySpeedPercentValidator);
             break;
         }
@@ -1123,7 +1123,7 @@ double MainWindow::getCustomSpeedFactor() const
 void MainWindow::seek(int value, SkyConnectIntf::SeekMode seekMode) const noexcept
 {
     auto &skyConnectManager = SkyConnectManager::getInstance();
-    const double factor = static_cast<double>(value) / static_cast<double>(PositionSliderMax);
+    const double factor = static_cast<double>(value) / static_cast<double>(::PositionSliderMax);
     const std::int64_t totalDuration = Logbook::getInstance().getCurrentFlight().getTotalDurationMSec();
     auto timestamp = static_cast<std::int64_t>(std::round(factor * static_cast<double>(totalDuration)));
 
@@ -1168,7 +1168,7 @@ void MainWindow::onTimeStampTimeEditChanged(const QTime &time) noexcept
 {
     auto &skyConnectManager = SkyConnectManager::getInstance();
     if (skyConnectManager.isIdle() || skyConnectManager.getState() == Connect::State::ReplayPaused) {
-        std::int64_t timestamp = time.hour() * MilliSecondsPerHour + time.minute() * MilliSecondsPerMinute + time.second() * MilliSecondsPerSecond;
+        std::int64_t timestamp = time.hour() * ::MilliSecondsPerHour + time.minute() * ::MilliSecondsPerMinute + time.second() * ::MilliSecondsPerSecond;
         skyConnectManager.seek(timestamp, SkyConnectIntf::SeekMode::Discrete);
     }
 }
@@ -1200,7 +1200,7 @@ void MainWindow::onTimestampChanged(std::int64_t timestamp) noexcept
 
 void MainWindow::onReplaySpeedSelected(QAction *action) noexcept
 {
-    ReplaySpeed replaySpeed = static_cast<ReplaySpeed>(action->property(ReplaySpeedProperty).toInt());
+    ReplaySpeed replaySpeed = static_cast<ReplaySpeed>(action->property(::ReplaySpeedProperty).toInt());
     float replaySpeedFactor {1.0f};
     switch (replaySpeed) {
     case ReplaySpeed::Slow10:
@@ -1344,7 +1344,7 @@ void MainWindow::updateControlUi() noexcept
         ui->skipToEndAction->setEnabled(false);
         // Position
         ui->positionSlider->setEnabled(false);
-        ui->positionSlider->setValue(PositionSliderMax);
+        ui->positionSlider->setValue(::PositionSliderMax);
         ui->timestampTimeEdit->setEnabled(false);
         break;
     case Connect::State::RecordingPaused: